pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "network-forensics-evidence-correlator"
version = "0.1.0"
description = "Correlates PCAP, firewall, and IDS evidence into a unified cyber-attack timeline."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "flask==3.1.2",
  "python-dateutil==2.9.0.post0",
  "tabulate==0.9.0",
  "waitress==3.0.2",
]

[tool.setuptools]
packages = ["src", "src.correlator"]
